+++ CHANGELOG.md
@@ -49,6 +49,7 @@ d'un utilisateur. L'affichage de cette nouvelle page a été optimisé. #18 ~Adm
 - Mise en place du nouveau schéma Actes V2.2 #177 ~Actes
 - Gestion du multi-canal : transmission complémentaire au format papier (ajout d'une case à cocher dans le formulaire de création d'un actes) #184 ~Actes 
 - Gestion multi-canal : information reprise dans la description de l'acte et dans le bordereau d'acquittement #184 ~Actes
+- Gestion de la typologie des pièces jointes #179 ~Actes
 
 ### Corrections